fix: remaining warnings generated by changes in Swift 6 with proposal SE-0364.
